backend/requirements.txt
aiohttp==3.5.4
aioredis==1.2.0
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
hiredis==1.0.0
idna==2.8
idna-ssl==1.1.0
jsonschema==2.6.0
multidict==4.5.2
python-engineio==3.2.3
python-socketio==3.1.1
six==1.12.0
typing-extensions==3.7.2
yarl==1.3.0
